[Decomp][Group] Cohesion split: extract instance binds to GroupInstance.cpp
